ml/requirements.txt
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
matplotlib==3.8.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
